+++ src/ClientCore/Rendering/Sprites/AnimatedSprites/AnimatedSpriteManager.cs
@@ -21,11 +21,13 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
+using Castle.Core.Logging;
+using Sovereign.EngineCore.Logging;
+using Sovereign.EngineCore.Main;
+using Sovereign.EngineCore.Resources;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Sovereign.ClientCore.Rendering.Sprites.AnimatedSprites
 {
@@ -36,14 +38,46 @@ namespace Sovereign.ClientCore.Rendering.Sprites.AnimatedSprites
     public sealed class AnimatedSpriteManager
     {
 
+        public ILogger Logger { private get; set; } = NullLogger.Instance;
+
+        public IErrorHandler ErrorHandler { private get; set; } = NullErrorHandler.Instance;
+
+        /// <summary>
+        /// Resource category used by the animated sprite definitions.
+        /// </summary>
+        public const ResourceType DefinitionsResourceType = ResourceType.Sprite;
+
+        /// <summary>
+        /// Animated sprite definitions filename.
+        /// </summary>
+        public const string DefinitionsFilename = "AnimatedSpriteDefinitions.yaml";
+
+        /// <summary>
+        /// Animated sprites.
+        /// </summary>
+        public readonly IList<AnimatedSprite> AnimatedSprites = new List<AnimatedSprite>();
+
         /// <summary>
         /// Animated sprite definitions loader.
         /// </summary>
         private readonly AnimatedSpriteDefinitionsLoader loader;
 
-        public AnimatedSpriteManager(AnimatedSpriteDefinitionsLoader loader)
+        /// <summary>
+        /// Resource path builder.
+        /// </summary>
+        private readonly IResourcePathBuilder pathBuilder;
+
+        /// <summary>
+        /// Sprite manager.
+        /// </summary>
+        private readonly SpriteManager spriteManager;
+
+        public AnimatedSpriteManager(AnimatedSpriteDefinitionsLoader loader,
+            IResourcePathBuilder pathBuilder, SpriteManager spriteManager)
         {
             this.loader = loader;
+            this.pathBuilder = pathBuilder;
+            this.spriteManager = spriteManager;
         }
 
         /// <summary>
@@ -51,7 +85,45 @@ namespace Sovereign.ClientCore.Rendering.Sprites.AnimatedSprites
         /// </summary>
         public void InitializeAnimatedSprites()
         {
+            var definitions = LoadDefinitions();
+            UnpackDefinitions(definitions);
+        }
 
+        /// <summary>
+        /// Loads the animated sprite definitions.
+        /// </summary>
+        /// <returns>Animated sprite definitions.</returns>
+        /// <exception cref="FatalErrorException">
+        /// Thrown if the definitions cannot be loaded.
+        /// </exception>
+        private AnimatedSpriteDefinitions LoadDefinitions()
+        {
+            var filename = pathBuilder.BuildPathToResource(DefinitionsResourceType, 
+                DefinitionsFilename);
+            try
+            {
+                return loader.LoadDefinitions(filename);
+            }
+            catch (Exception e)
+            {
+                /* Log and throw a fatal error. */
+                Logger.Fatal("Failed to load animated sprite definitions.", e);
+                ErrorHandler.Error(e.Message);
+
+                throw new FatalErrorException();
+            }
+        }
+
+        /// <summary>
+        /// Unpacks the animated sprite definitions.
+        /// </summary>
+        /// <param name="definitions">Definitions to unpack.</param>
+        private void UnpackDefinitions(AnimatedSpriteDefinitions definitions)
+        {
+            foreach (var def in definitions.AnimatedSprites.OrderBy(def => def.Id))
+            {
+                AnimatedSprites.Add(new AnimatedSprite(def, spriteManager));
+            }
         }
 
     }